pyproject.toml
[project]
name = "sidantrip"
version = "0.1.0"
description = "SidanTrip AI agent service — CrewAI + LiteLLM travel planner"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[anthropic,tools]>=1.9.3",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sse-starlette>=2.0.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "black"]

[project.scripts]
sidantrip = "sidantrip.main:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
